// sanity.types.ts
export type Slug = {
  _type: 'slug';
  current?: string;
  source?: string;
};
// ...
export type CombinedSearchQueryResult = {
  blogs: Array<{
    _id: string;
    title: string | null;
    slug: Slug | null;
  }>;
  projects: Array<{
    _id: string;
    title: string | null;
    slug: Slug | null;
  }>;
  categories: Array<{
    _id: string;
    name: string | null;
    slug: Slug | null;
  }>;
  tags: Array<{
    _id: string;
    name: string | null;
    slug: Slug | null;
  }>;
};